requirements.txt
anaconda==0.0.1.1
backcall==0.1.0
beautifulsoup4==4.6.3
bs4==0.0.1
cassandra-driver==3.13.0
certifi==2018.8.24
cfblog==0.2.16
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.8
django-allauth==0.35.0
django-cassandra-engine==1.5.3
django-extensions==2.1.3
django-filter==2.0.0
django-follow==0.6.1
django-generic-follow==0.4.2
django-language-field==0.0.3
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-redis==4.9.0
django-tagging==0.4.6
django-timezone-field==3.0rc1
django-versatileimagefield==1.9
djangorestframework==3.8.2
feedparser==5.2.1
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
jsonfield==2.0.2
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0.1
matplotlib==3.0.0
mistune==0.8.4
numpy==1.15.2
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.0.0
profanity==1.1
prompt-toolkit==2.0.5
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
selenium==3.14.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
wikipedia==1.4.0
wordcloud==1.5.0
youtube==1.0.6
youtube-python==1.0.12
